Converted the XGBoostmodule to contain a model class. This is to keep with better pracitises in regards to OOP programming.
There is a subsampling routine in the model class which splits the data into a training and testing set. This is so that the dataset is loaded once and then randomly chosen to in the same way each time.

The Datacuts have been added to the Main module.
