+++ Particle_DataSet/081220222253.csv
@@ -1,0 +1,2 @@
+round_epochs,val_loss,val_acc,loss,acc,activation,emb_output_dims,epochs,first_neuron,hidden_layers,last_activation,losses,lr,optimizer,shape,weight_regulizer
+5,0.0007864532337233537,0.9998071789741516,0.0008270713670423219,0.9997829,relu,None,5,4,0,sigmoid,binary_crossentropy,0.5,Adam,brick,None
